pull in common submodule
The common submodule includes openshift-auth-proxy for the convenience
of tagging and releasing from a single repo.

It also contains genericized elasticsearch and curator definitions which
will replace the specialized ones in the near future.

It also includes test and deployer/validation utilities which should
be used in the deployer eventually (which is why the submodule is
placed under the deployer).